refactor: ModalHeader 컴포넌트로 파일 분리
