+++ src/pages/hj.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import SideBar from '@/components/common/DashBoardHeader/SideBar';
 import DashBoardHeader from '@/components/common/DashBoardHeader/index';
-import CardConfirmModal from '@/components/common/Modal/CardConfirmModal';
+import CardConfirmModal from '@/components/common/Modal/CardConfirmModal/CardConfirmModal';
 import HashTag from '@/components/common/tag/HashTag';
 import StateTag from '@/components/common/tag/StateTag';
 import '@/styles/GlobalStyle';
